requirements.txt
base58
python-dotenv
python-socketio==5.11.0
redis
uvicorn
